BiodiversityObservationNetworks.jl

The purpose of this package is to provide a high-level, extensible, modular interface to the selection of sampling point for biodiversity processes in space. It is based around a collection of types representing point selection algorithms, used to select the most informative sampling points based on raster data. Specifically, the algorithms work from a layer indicating entropy at each location.

This package is in development

The BiodiversityObservationNetworks.jl package is currently under development. The API is not expected to change a lot, but it may change in order to facilitate the integration of new features.

High-level types

BiodiversityObservationNetworks.BONSampler — Type

BONSampler

A union of the abstract types BONSeeder and BONRefiner. Both types return a tuple with the coordinates as a vector of CartesianIndex, and the weight matrix as a Matrix of AbstractFloat, in that order.

BiodiversityObservationNetworks.BONSeeder — Type

abstract type BONSeeder end

A BONSeeder is an algorithm for proposing sampling locations using a raster of weights, represented as a matrix, in each cell.

BiodiversityObservationNetworks.BONRefiner — Type

abstract type BONRefiner end

A BONRefiner is an algorithm for proposing sampling locations by refining a set of candidate points to a smaller set of 'best' points.

Seeder and refiner functions

BiodiversityObservationNetworks.seed — Function

seed(sampler::ST, uncertainty::Matrix{T})

Produces a set of candidate sampling locations in a vector coords of length numpoints from a raster uncertainty using sampler, where sampler is a BONSeeder.

seed!(coords::Vector{CartesianIndex}, sampler::ST)

The curried version of seed!, which returns a function that acts on the input uncertainty layer passed to the curried function (u below).

BiodiversityObservationNetworks.seed! — Function

seed!(coords::Vector{CartesianIndex}, sampler::ST, uncertainty::Matrix{T})

Puts a set of candidate sampling locations in the preallocated vector coords from a raster uncertainty using sampler, where sampler is a BONSeeder.

Seeder's work on rasters, refiners work on set of coordinates.

seed!(coords::Vector{CartesianIndex}, sampler::ST)

The curried version of seed!, which returns a function that acts on the input uncertainty layer passed to the curried function (u below).

BiodiversityObservationNetworks.refine — Function

refine(pool::Vector{CartesianIndex}, sampler::ST, uncertainty::Matrix{T})

Refines a set of candidate sampling locations and returns a vector coords of length numpoints from a vector of coordinates pool using sampler, where sampler is a BONRefiner.

refine(sampler::BONRefiner)

Returns a curried function of refine with two methods: both are using the output of seed, one in its packed form, the other in its splatted form.

refine(pack, sampler::BONRefiner)

Calls refine on the appropriatedly splatted version of pack.

BiodiversityObservationNetworks.refine! — Function

refine!(cooords::Vector{CartesianIndex}, pool::Vector{CartesianIndex}, sampler::ST, uncertainty::Matrix{T})

Refines a set of candidate sampling locations in the preallocated vector coords from a vector of coordinates pool using sampler, where sampler is a BONRefiner.

refine!(cooords::Vector{CartesianIndex}, pool::Vector{CartesianIndex}, sampler::ST, uncertainty::Matrix{T})

The curried version of refine!, which returns a function that acts on the input coordinate pool passed to the curried function (p below).

Seeder algorithms

BiodiversityObservationNetworks.BalancedAcceptance — Type

BalancedAcceptance

A BONSeeder that uses Balanced-Acceptance Sampling (Van-dem-Bates et al. 2017 https://doi.org/10.1111/2041-210X.13003)

Refiner algorithms

BiodiversityObservationNetworks.AdaptiveSpatial — Type

AdaptiveSpatial

...

numpoints, an Integer (def. 50), specifying the number of points to use.

α, an AbstractFloat (def. 1.0), specifying ...

Helper functions

BiodiversityObservationNetworks.entropize! — Function

entropize!(U::Matrix{AbstractFloat}, A::Matrix{Number})

This function turns a matrix A (storing measurement values) into pixel-wise entropy values, stored in a matrix U (that is previously allocated).

Pixel-wise entropy is determined by measuring the empirical probability of randomly picking a value in the matrix that is either lower or higher than the pixel value. The entropy of both these probabilities are calculated using the -p×log(2,p) formula. The entropy of the pixel is the sum of the two entropies, so that it is close to 1 for values close to the median, and close to 0 for values close to the extreme of the distribution.

BiodiversityObservationNetworks.entropize — Function

entropize(A::Matrix{Number})

Allocation version of entropize!.